Реализован CConfig. Является синглтоном и позволяет легко прочитать или записать настройки в любом месте. Гибкая структура классов-бэкендов позволит создавать побочные конфиги помимо главного. Сделан класс CEnvironment::Variables с функциями Get и Set, позволяющими манипулировать переменными среды. Теперь путь к логам берётся из CEnvironment::Paths::GetLogPath. Соответствующая функция CEnvironment::Paths::SetLogPath позволяет этот путь устанавливать. Определение пути к исполняемому файлу (windows-only) вынесено в CEnvironment::Paths::GetExecutablePath. А CEnvironment::Paths::SetWorkingDirectory теперь действительно может изменять working directory. Добавлено понятие универсальной директории - см. CEnvironment::Paths::GetUniversalDirectory. Может использоваться для кросс-платформенного определения места, где можно хранить изменяемые файлы. В XML добавлены функции CXMLChildrenList::Remove(CXMLNode *ANode), CXMLNode[WithAttributes]::HasAttribute. CVariant удалён, так как не смог обеспечить должной гибкости. Вместо него теперь есть CVariantConvert, который с одной стороны проще, с другой - удобнее. Также теперь есть глобальная шаблонная функция from_string, противоположность к to_string. В качестве временного решения до введения системы ивентов реализовано разделение кей-ивентов, посылаемых в InputHandling, на KEY_DOWN, KEY_PRESS и KEY_UP. KEY_PRESS подвержен KeyRepeat'у, остальные - нет. Пофикшен билд под 64-битными системами (были кое-какие проблемы с размерами целочисленных переменных). Изменения в Makefile'ах, в частности для лучшей совместимости. Удалён Engine/doc/paths.pdf, так как более не актуален.
Update issue #18
Status: Fixed
CConfig сделан. Осталось реализовать сопоставление конфига с аргуменатми командной строки, но это будет отдельной иссуей.

+++ Engine/src/2de_GameUtils.cpp
@@ -246,8 +246,8 @@ int CLevelMap::GetCellIndex(int h, int v)
 
 CBox CLevelMap::GetCellAABB(const Vector2 &V)
 {
-	size_t CellWidth = TileSet->TileWidth * GetScaling();
-	size_t CellHeight = TileSet->TileWidth * GetScaling();
+	unsigned int CellWidth = TileSet->TileWidth * GetScaling();
+	unsigned int CellHeight = TileSet->TileWidth * GetScaling();
 	if (GetMapCell((int)V.x / CellWidth, (int)V.y / CellHeight)->index != 0)
 		return CBox((((int)V.x / CellWidth)) * CellWidth,
 		(((int)V.y/CellHeight)) * CellHeight,
